requirements.txt
anyio==4.11.0
asgiref==3.10.0
black==25.9.0
blinker==1.9.0
build==1.3.0
CacheControl==0.14.3
certifi==2025.10.5
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.4.0
Django==5.2.7
djangorestframework==3.16.1
dulwich==0.24.6
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.0
flake8==7.3.0
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
keyring==25.6.0
MarkupSafe==3.0.3
mccabe==0.7.0
more-itertools==10.8.0
msgpack==1.1.2
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pbs-installer==2025.10.14
pipenv==2025.0.4
pkginfo==1.12.1.2
platformdirs==4.5.0
pluggy==1.6.0
poetry==2.2.1
poetry-core==2.2.1
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-django==4.11.1
pytokens==0.2.0
pywin32-ctypes==0.2.3
RapidFuzz==3.14.1
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tomlkit==0.13.3
trove-classifiers==2025.9.11.17
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.3
Werkzeug==3.1.3
zstandard==0.25.0
